requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-filter==2.4.0
whitenoise==5.0.1
gunicorn==19.9.0
html5lib==1.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
msgpack==1.0.2
Pillow==8.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
zope.interface==5.4.0
